EPI-1756: Added mock logic to progress fills
- Added mock logic to simulate Endur side deal booking and to provide
arbitrary deal tracking nums once a fill is being posted
- Made the tradeId field in the fill object optional
- Adapted test cases and other parts of the logic accordingly

+++ TOMS/Backend/TOMS-MockService/src/main/java/com/matthey/pmm/toms/service/mock/OrderTestDataGenerator.java
@@ -76,7 +76,7 @@ public class OrderTestDataGenerator {
 
 	private static final double MAX_REFERENCE_ORDER_LEG_NOTIONAL = 100;
 
-	private static int FILL_TRADE_ID_COUNTER=10000000;
+	private static long FILL_TRADE_ID_COUNTER=10000000;
 
 	
 	@Autowired
